refactor: jogoLuta lib file structure
