+++ Jogo Luta/src/js/jogoLuta/Controller.js
