+++ Jogo Luta/src/js/jogoLuta/Logger.js
