+++ Jogo Luta/src/js/jogoLuta/Stage.js
