+++ Jogo Luta/src/js/jogoLuta/characters/Knight.js
