+++ Jogo Luta/src/js/jogoLuta/characters/LittleMonster.js
