+++ Jogo Luta/src/js/jogoLuta/characters/Sorcerer.js
