+++ Jogo Luta/src/js/jogoLuta/characters/index.js
